+++ framework/model/src/main/java/is/codion/framework/model/EntityConditionModel.java
@@ -21,7 +21,7 @@ package is.codion.framework.model;
 import is.codion.common.Conjunction;
 import is.codion.common.model.condition.ConditionModel;
 import is.codion.common.model.condition.TableConditionModel;
-import is.codion.common.value.Value;
+import is.codion.common.observer.Mutable;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.Attribute;
@@ -70,32 +70,34 @@ public interface EntityConditionModel extends TableConditionModel<Attribute<?>> 
 	<T> boolean setInOperands(Attribute<T> attribute, Collection<T> operands);
 
 	/**
-	 * Returns a WHERE condition based on enabled condition models which are based on non-aggregate function columns.
+	 * Returns a WHERE condition based on enabled condition models which are based on non-aggregate function columns
+	 * along with any {@link #additionalWhere()} condition.
 	 * @param conjunction the conjunction to use in case of multiple enabled conditions
-	 * @return the current where condition based on the state of the underlying condition models
+	 * @return the current WHERE condition based on the state of the underlying condition models
 	 */
 	Condition where(Conjunction conjunction);
 
 	/**
-	 * Returns a HAVING condition based on enabled condition models which are based on aggregate function columns.
+	 * Returns a HAVING condition based on enabled condition models which are based on aggregate function columns
+	 * along with any {@link #additionalHaving()} condition.
 	 * @param conjunction the conjunction to use in case of multiple enabled conditions
-	 * @return the current having condition based on the state of the underlying condition models
+	 * @return the current HAVING condition based on the state of the underlying condition models
 	 */
 	Condition having(Conjunction conjunction);
 
 	/**
-	 * Controls the additional where condition.
+	 * Controls the additional WHERE condition.
 	 * The condition supplier may return null in case of no condition.
-	 * @return the {@link Value} controlling the additional where condition
+	 * @return the {@link AdditionalCondition} instance controlling the additional WHERE condition
 	 */
-	Value<Supplier<Condition>> additionalWhere();
+	AdditionalCondition additionalWhere();
 
 	/**
-	 * Controls the additional having condition.
+	 * Controls the additional HAVING condition.
 	 * The condition supplier may return null in case of no condition.
-	 * @return the {@link Value} controlling the additional having condition
+	 * @return the {@link AdditionalCondition} instance controlling the additional HAVING condition
 	 */
-	Value<Supplier<Condition>> additionalHaving();
+	AdditionalCondition additionalHaving();
 
 	/**
 	 * Returns the {@link ConditionModel} associated with the given attribute.
@@ -126,5 +128,17 @@ public interface EntityConditionModel extends TableConditionModel<Attribute<?>> 
 	static EntityConditionModel entityConditionModel(EntityType entityType, EntityConnectionProvider connectionProvider,
 																									 ConditionModel.Factory<Attribute<?>> conditionModelFactory) {
 		return new DefaultEntityConditionModel(entityType, connectionProvider, conditionModelFactory);
+	}
+
+	/**
+	 * Specifies an additional condition supplier.
+	 */
+	interface AdditionalCondition extends Mutable<Supplier<Condition>> {
+
+		/**
+		 * Default {@link Conjunction#AND}.
+		 * @return the {@link Mutable} controlling the {@link Conjunction} to use when adding the additional condition
+		 */
+		Mutable<Conjunction> conjunction();
 	}
 }